move components to the base theme, render usermenu in course theme as well
