+++ course-v2/assets/course-v2.tsx
@@ -12,6 +12,11 @@ import {
   checkAnswer,
   showSolution
 } from "./js/quiz_multiple_choice"
+import { QueryClientProvider } from "@tanstack/react-query"
+import React from "react"
+import ReactDOM from "react-dom/client"
+import { makeQueryClient } from "../../base-theme/assets/js/clients"
+import UserMenu from "../../base-theme/assets/js/components/UserMenu"
 
 export interface OCWWindow extends Window {
   initNanogallery2: () => void
@@ -27,12 +32,18 @@ $(function() {
   checkAnswer()
   showSolution()
   initCourseDrawersClosingViaSwiping()
+  const userMenuContainer = document.querySelector("#user-menu-container")
+  if (userMenuContainer) {
+    const queryClient = makeQueryClient()
+    const root = ReactDOM.createRoot(userMenuContainer)
+    root.render(<QueryClientProvider client={queryClient}><UserMenu /></QueryClientProvider>)
+  }
 })
 
 let nanogallery2Loaded = false
 
 window.initNanogallery2 = () => {
   if (nanogallery2Loaded) return
-  import("./nanogallery2-imports")
+  import("./nanogallery2-imports.js")
   nanogallery2Loaded = true
 }
