+++ www/assets/www.tsx
@@ -4,7 +4,7 @@ import "./css/search.scss"
 import Popper from "popper.js"
 import React from "react"
 import ReactDOM from "react-dom/client"
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import { QueryClientProvider } from "@tanstack/react-query"
 import { createBrowserHistory } from "history"
 
 import SearchPage from "./js/components/SearchPage"
@@ -14,7 +14,8 @@ import { setupEmailSignupForm } from "./js/mailchimp"
 import { initNotifications } from "./js/notification"
 import { initSubNav } from "./js/subnav"
 import ResourceCollection from "./js/components/ResourceCollection"
-import UserMenu from "./js/components/UserMenu"
+import UserMenu from "../../base-theme/assets/js/components/UserMenu"
+import { makeQueryClient } from "../../base-theme/assets/js/clients"
 
 export interface OCWWindow extends Window {
   $: JQueryStatic
@@ -29,43 +30,6 @@ window.$ = $
 window.Popper = Popper
 
 const history = createBrowserHistory()
-
-type MaybeHasStatus = {
-  response?: {
-    status?: number
-  }
-}
-
-const RETRY_STATUS_CODES = [408, 429, 502, 503, 504]
-const MAX_RETRIES = 3
-const THROW_ERROR_CODES: (number | undefined)[] = [404, 403, 401]
-
-const makeQueryClient = (): QueryClient => {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        refetchOnWindowFocus: false,
-        staleTime: Infinity,
-        throwOnError: (error) => {
-          const status = (error as MaybeHasStatus)?.response?.status
-          return THROW_ERROR_CODES.includes(status)
-        },
-        retry: (failureCount, error) => {
-          const status = (error as MaybeHasStatus)?.response?.status
-          /**
-           * React Query's default behavior is to retry all failed queries 3
-           * times. Many things (e.g., 403, 404) are not worth retrying. Let's
-           * just retry some explicit whitelist of status codes.
-           */
-          if (status !== undefined && RETRY_STATUS_CODES.includes(status)) {
-            return failureCount < MAX_RETRIES
-          }
-          return false
-        },
-      },
-    },
-  })
-}
 
 $(function() {
   const userMenuContainer = document.querySelector("#user-menu-container")
